% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mgx_model.R
\name{fit_mgx_model}
\alias{fit_mgx_model}
\title{Fit model for fixed gene k, taxon j, taxon j*}
\usage{
fit_mgx_model(
  yy,
  xstar,
  xx,
  replicates = NULL,
  formula = NULL,
  enviro_df = NULL,
  wts = NULL,
  replace_zeros = "minimum",
  use_jack_se = FALSE,
  cluster_corr_coef = NULL
)
}
\arguments{
\item{yy}{vector of abundances of gene k in taxon j}

\item{xstar}{vector of abundances of taxon j*}

\item{xx}{vector of abundances of taxon j}

\item{replicates}{vector that contains the same entries for all observations that are technical replicates. Will be coerced to a factor down the line.}

\item{formula}{a formula describing the environmental covariates that should be included in the model. The variable names should be columns in enviro_df}

\item{enviro_df}{a data frame or tibble with columns containing environmental covariates that should be included in the model,}

\item{wts}{vector of nonnegative weights. Could be sequencing depth to put more emphasis on deeply sequenced samples}

\item{replace_zeros}{what to do with zeros in the denominator \eqn{X_{ij}}. Options include "minimum" or pseudocount numeric value (eg. 1)}

\item{use_jack_se}{when replicates are given, if TRUE will use jackknife standard errors instead of sandwich standard errors. This is recommended when
there is a small number of clusters.}

\item{cluster_corr_coef}{when replicates are given, estimated value of the within-cluster correlation coefficient. This will only be used when gee estimation in \code{raoBust::gee_test} fails, and instead
estimation is performed with a glm. This is set to NULL by default.}
}
\description{
We would like to estimate beta1 in the model
\eqn{Y_ijk ~ Poisson(X_ij * c * (X_{ij*} / X_{ij})^{beta_1})}
where
\eqn{Y_{ijk}} refers to data about the abundance of gene k expressed by taxon j in sample i
\eqn{X_{ij}} refers to data about the abundance of taxon j in sample i
\eqn{X_{ij*}} refers to data about the abundance of taxon j* in sample i
}
\details{
We can then interpret, e.g., \eqn{1.01^{beta_1}} as the multiplicative change in the expression-per-unit-coverage of gene k in species j for 1\\% increase in the coverage of species j compared to species j*. With some slightly stronger assumptions about the sampling mechanism, we can also interpret this on the abundance (rather than just coverage) scale
}
