% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mgx_model.R
\name{fit_mgx_model}
\alias{fit_mgx_model}
\title{Fit model for fixed gene k, taxon j, taxon j*}
\usage{
fit_mgx_model(
  enviro_df,
  yy = "yy",
  xstar = "xstar",
  xx = "xx",
  replicates = NULL,
  formula = NULL,
  wts = NULL,
  replace_zeros = "minimum",
  use_jack_se = FALSE,
  cluster_corr_coef = NULL
)
}
\arguments{
\item{enviro_df}{a data frame or tibble with columns containing relevant abundances, environmental covariates that should be included in the model, and optionally replicate information}

\item{yy}{column of \code{enviro_df} containing abundances of gene k in taxon j. Default is "yy".}

\item{xstar}{column of \code{enviro_df} containing abundances of taxon j*. Default is "xstar".}

\item{xx}{column of \code{enviro_df} containing abundances of taxon j. Default is "xx".}

\item{replicates}{column of \code{enviro_df} containing the same entries for all observations that are technical replicates. Will be coerced to a factor down the line. Default is \code{NULL}.}

\item{formula}{a formula describing the environmental covariates that should be included in the model. The variable names should be columns in \code{enviro_df}}

\item{wts}{column of \code{enviro_df} containing nonnegative weights. Could be sequencing depth to put more emphasis on deeply sequenced samples. Default is \code{NULL}.}

\item{replace_zeros}{what to do with zeros in the denominator \eqn{X_{ij}}. Options include "minimum" or pseudocount numeric value (eg. 1)}

\item{use_jack_se}{when replicates are given, if TRUE will use jackknife standard errors instead of sandwich standard errors. This is recommended when
there is a small number of clusters.}

\item{cluster_corr_coef}{when replicates are given, estimated value of the within-cluster correlation coefficient. This will only be used when gee estimation in \code{raoBust::gee_test} fails, and instead
estimation is performed with a glm. This is set to NULL by default.}
}
\description{
We would like to estimate beta1 in the model
\eqn{Y_ijk ~ Poisson(X_ij * c * (X_{ij*} / X_{ij})^{beta_1})}
where
\eqn{Y_{ijk}} refers to data about the abundance of gene k expressed by taxon j in sample i
\eqn{X_{ij}} refers to data about the abundance of taxon j in sample i
\eqn{X_{ij*}} refers to data about the abundance of taxon j* in sample i
}
\details{
We can then interpret, e.g., \eqn{1.01^{beta_1}} as the multiplicative change in the expression-per-unit-coverage of gene k in species j for 1\\% increase in the coverage of species j compared to species j*. With some slightly stronger assumptions about the sampling mechanism, we can also interpret this on the abundance (rather than just coverage) scale
}
\examples{
my_df <- data.frame(xx = rpois(20, lambda = 400),
                    xstar = rpois(20, lambda = 400),
                    treatment = rep(0:1, each = 10),
                    id = rep(1:4, 5))
my_df$yy <- rpois(20, my_df$xx * 10 * (my_df$xstar / my_df$xx)^2)
fit_mgx_model(enviro_df = my_df, replicates = "id", formula = ~ treatment)

}
