% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mgx_model.R
\name{fit_mgx_model}
\alias{fit_mgx_model}
\title{Estimate the difference in expression-per-unit-abundance of a gene expressed by a responder taxon as a function of the abundance of a companion taxon}
\usage{
fit_mgx_model(
  enviro_df,
  yy = "yy",
  xstar = "xstar",
  xx = "xx",
  replicates = NULL,
  formula = NULL,
  wts = NULL,
  replace_zeros = "minimum",
  use_jack_se = FALSE,
  cluster_corr_coef = NULL
)
}
\arguments{
\item{enviro_df}{A data frame or tibble with columns containing the relevant variables, and rows denoting the observations. Columns must contain the following data: observed gene expression data, taxon abundance data, environmental covariates, and (optionally) technical replicate information.}

\item{yy}{The name of the column in \code{enviro_df} containing the expression data for a gene expressed by the responder taxon. Default is "yy". Expression data could be in the form of coverage, counts, etc. See vignettes for details on acceptable and suggested datatypes and preprocessing ("normalizations"). In the above notation, this corresponds to the expression of gene \eqn{k} expressed by taxon \eqn{j}.}

\item{xstar}{The name of the column in \code{enviro_df} containing abundances of the companion taxon. Default is "xstar". In the above notation, this corresponds to the abundance of  taxon \eqn{j^*}.}

\item{xx}{The name of the column in \code{enviro_df} containing abundances of the responder taxon. Default is "xx". In the above notation, this corresponds to the abundance of taxon \eqn{j}.}

\item{replicates}{The name of the column in \code{enviro_df} containing technical replicate information. Entries in this column should be the same within observations that are technical replicates. Will be coerced to a factor internally. Default is \code{NULL}.}

\item{formula}{a formula describing the environmental covariates that should be included in the model. The variable names should be columns in \code{enviro_df}}

\item{wts}{The name of the column in \code{enviro_df} containing nonnegative weights for the observation. Could be sequencing depth to put more emphasis on deeply sequenced samples. Default is \code{NULL}.}

\item{replace_zeros}{what to do with zeros in the denominator \eqn{X_{ij}}. Options include "minimum" (replace with the smallest value) or pseudocount numeric value (eg. 1). Default is "minimum".}

\item{use_jack_se}{Boolean indicating whether to use jackknife standard errors (TRUE) rather than sandwich standard errors (FALSE). Only relevant if technical replicates are included. Jackknife standard errors are recommended when
there are a small number of observations. Defaults to FALSE.}

\item{cluster_corr_coef}{When technical replicates are given, the estimated value of the within-cluster correlation coefficient. This will only be used when GEE estimation in \code{raoBust::gee_test} fails, and
estimation is performed with a glm. Defaults to NULL by default (no robust score test is returned).}
}
\description{
We would like to estimate the \eqn{\beta} parameters in the model
\eqn{\text{average } Y_{ijk} = X_{ij} \times c \times (X_{ij^*} / X_{ij})^{\beta_1} \times e^{\beta_2 W_{i1} + \beta_3 W_{i2} + \ldots + \beta_p W_{i,p-1}}},
where
\eqn{Y_{ijk}} refers to the observed abundance of gene \eqn{k} expressed by taxon \eqn{j} in sample \eqn{i},
\eqn{X_{ij}} refers to the observed abundance of taxon \eqn{j} in sample \eqn{i},
\eqn{X_{ij^*}} refers to the observed abundance of taxon \eqn{j^*} in sample \eqn{i}, and
\eqn{W_{im}} is the value of abiotic covariate \eqn{m} observed in sample \eqn{i}.
We can interpret \eqn{1.01^{\beta_1}} as the multiplicative change in the expression-per-unit-coverage of gene \eqn{k} in species \eqn{j} for 1\\\% increase in the coverage of species \eqn{j} compared to species \eqn{j^*}. With some slightly stronger assumptions about the sampling mechanism, we can also interpret this on the true cell abundance scale (rather than just the coverage scale).
}
\examples{
my_df <- data.frame(xx = rpois(20, lambda = 400),
                    xstar = rpois(20, lambda = 400),
                    treatment = rep(0:1, each = 10),
                    id = rep(1:4, 5))
my_df$yy <- rpois(20, my_df$xx * 10 * (my_df$xstar / my_df$xx)^2)
fit_mgx_model(enviro_df = my_df, replicates = "id", formula = ~ treatment)

}
