% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mgx_model.R
\name{fit_mgx_model}
\alias{fit_mgx_model}
\title{Fit model for fixed gene k, taxon j, taxon j*}
\usage{
fit_mgx_model(yy, xstar, xx, wts = NULL, replace_zeros = "minimum")
}
\arguments{
\item{yy}{vector of abundances of gene k in taxon j}

\item{xstar}{vector of abundances of taxon j*}

\item{xx}{vector of abundances of taxon j}

\item{wts}{vector of nonnegative weights. Could be sequencing depth to put more emphasis on deeply sequenced samples}

\item{replace_zeros}{what to do with zeros in the denominator X_{ij}. Options include "minimum" or pseudocount numeric value (eg. 1)}
}
\description{
Fit model for fixed gene k, taxon j, taxon j*
}
